requirements.txt
gitpython
requests
keyring
docopt
